Move ConstScorer to const_score_query module, export ConstScorQuery and ConstScoreQuery from parent query module
